requirements.txt
tensorflow==2.4.1
Pillow==8.4.0
